Refactor codebase into separate directories
