+++ src/main/java/duke/internals/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.internals;
 
 import duke.command.ByeCommand;
 import duke.command.Command;
