+++ src/main/java/duke/ui/DialogBox.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 import java.io.IOException;
 import java.util.Collections;
 
@@ -16,6 +18,7 @@ import javafx.scene.layout.HBox;
  * An example of a custom control using FXML.
  * This control represents a dialog box consisting of an ImageView to represent the speaker's face and a label
  * containing text from the speaker.
+ * Source: https://se-education.org/guides/tutorials/javaFxPart4.html
  */
 public class DialogBox extends HBox {
     @FXML
