+++ src/main/java/duke/ui/Launcher.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 import javafx.application.Application;
 
 /**
